+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.3.0] - 2026-04-23
+- Updated the SDK models and parsing to match the current API contract.
+- Rebranded SDK docs to DSX Air.
+- Renamed disable_auto_oob_dhcp to enable_dhcp and "attributes" to "labels" for the Node and Interface endpoints.
+- Updated SDK Jupyter Notebook examples.
+- Added org_id and ngc_org_name fields to the Organization dataclass given recent API changes.
+
 # [1.2.0] - 2026-03-25
 - Added `Checkpoint` model and `CheckpointEndpointAPI` to the SDK, enabling users to list, retrieve, update, and delete simulation checkpoints.
 - Changed stubs of delete method in `ServiceEndpointAPI` to accept service ID only